pa/backend/utils/helpers.py
from bson import ObjectId
from datetime import datetime


def serialize_doc(doc):
    if doc is None:
        return None
    result = {}
    for key, value in doc.items():
        if key == "_id":
            result["id"] = str(value)
        elif isinstance(value, ObjectId):
            result[key] = str(value)
        elif isinstance(value, datetime):
            result[key] = value.isoformat()
        elif isinstance(value, list):
            result[key] = [serialize_doc(item) if isinstance(item, dict) else str(item) if isinstance(item, ObjectId) else item for item in value]
        elif isinstance(value, dict):
            result[key] = serialize_doc(value)
        else:
            result[key] = value
    return result


def calculate_production_percentage(production_value: float, target_value: float) -> float:
    if not target_value or target_value == 0:
        return 0.0
    return round((production_value / target_value) * 100, 2)
