pa/backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
motor==3.3.2
pymongo==4.6.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
pydantic[email]==2.6.4
openpyxl==3.1.2
pandas==2.2.1
python-dotenv==1.0.1
